pyproject.toml
[project]
name = "crypto-analytical-pipeline"
version = "0.6.0"
description = "A robust, orchestrated pipeline for downloading and analyzing Binance market data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "natsort>=8.4.0",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "duckdb>=0.10.0",
    "pyyaml>=6.0",
    "prefect>=2.16.0",
    "dbt-duckdb>=1.7.0"
]

[project.scripts]
crypto-pipeline = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crypto_pipeline"]
